implement a function "slice.ContainsStringEqualFold"
This function can be usefull and more easy to use instead of "slice.ContainsString"
becasue is not necesary the user to provide modifier function. Also if the user
wants to compare a string to the items of a slice with case insensitive, he
can't rely to pass a modifier to "slice.ContainsString(...)", becasue modifier
is applied only to the element of the slice ("aA" and "Aa" will be
always not equal).